+++ .gitignore
@@ -1311,6 +1311,7 @@ packages/lib/services/ocr/drivers/OcrDriverTesseract.js
 packages/lib/services/ocr/utils/filterOcrText.test.js
 packages/lib/services/ocr/utils/filterOcrText.js
 packages/lib/services/ocr/utils/getPageSize.js
+packages/lib/services/ocr/utils/htmlOverlayGenerator.js
 packages/lib/services/ocr/utils/types.js
 packages/lib/services/plugins/BasePlatformImplementation.js
 packages/lib/services/plugins/BasePluginRunner.js
